Relocate/rename hasqdivltcg variable as has_qdiv_or_ltcg
