+++ policyengine_us/tools/taxcalc/variable_mapping.yaml
@@ -53,7 +53,7 @@ tax_unit_weight: s006
 business_is_qualified: pt_sstb_income
 w2_wages_from_qualified_business: pt_binc_w2_wages
 unadjusted_basis_qualified_property: pt_ubia_property
-hasqdivltcg: hasqdivltcg
+has_qdiv_or_ltcg: hasqdivltcg
 medical_expense_deduction: c17000
 salt_deduction: c18300
 interest_deduction: c19200
