+++ policyengine_us/variables/gov/irs/tax/federal_income/capital_gains/dwks10.py
@@ -25,7 +25,7 @@ class dwks10(Variable):
             ),
         ) + add(tax_unit, period, ["non_sch_d_capital_gains"])
         return where(
-            tax_unit("hasqdivltcg", period),
+            tax_unit("has_qdiv_or_ltcg", period),
             dwks10_if_gains,
             dwks10_if_no_gains,
         )
